more warnings, and fixes

+++ main.c
@@ -10,7 +10,7 @@
 
 int main (int argc, char **argv) {
     int display_precision = (int)strtol(argv[1], NULL, BASE);
-    double internal_precision = strtod(argv[2], NULL) * 3.322;
+    long double internal_precision = strtod(argv[2], NULL) * 3.322L;
     mpfr_set_default_prec((int)internal_precision);
     fprintf(stderr, " MPFR default precision: %lu bits\n", mpfr_get_default_prec());
     t_init(display_precision);
